Organized files correctly.  Added watch:all to pkg.json.  Connected to Heroku. #3
