+++ package.json
@@ -1,12 +1,18 @@
 {
   "name": "backupband",
-  "version": "1.0.0",
-  "description": "A web app for JS developers want to save notes about what they've learned.",
+  "version": "0.3.0",
+  "engines": {
+    "node": "8.x",
+  "description": "A web app for songwriters to quickly and easily add backing music to their songs.",
   "main": "app.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "bundle": "browserify client/index.js -o server/public/main.js -t babelify",
-    "watchbundle": "watchify client/index.js -o server/public/main.js -t babelify"
+    "watch:client": "watchify client/index.js -o server/public/main.js -t babelify",
+    "start": "node server/express.js",
+    "watch:server": "nodemon server/express.js",
+    "watch:all": "run-p watch:client watch:server",
+    "postinstall": "npm run bundle"
   },
   "repository": {
     "type": "git",
@@ -20,6 +26,7 @@
     "babel-plugin-transform-react-jsx": "^6.24.1",
     "babelify": "^7.3.0",
     "browserify": "^14.4.0",
+    "npm-run-all": "^4.0.2",
     "stylelint": "^8.0.0",
     "stylelint-config-standard": "^17.0.0",
     "watchify": "^3.9.0"
