+++ server/public/pics/kick-hihat.png
